requirements.txt
Flask>=2.3
python-dotenv>=1.0
openai>=1.0
rdflib>=6.1
unittest-xml-reporting>=3.2
